services/worker-py/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
requests==2.32.4
python-dotenv==1.1.1
pytest==8.4.2
